Subida de CrudUsuarioBueno

+++ RA3BUENOOOO/EjemploLombokCOMPLETO/src/main/java/com/example/EjemploLombok/Repository/UsuarioRepository.java
@@ -14,7 +14,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     boolean existsByUsername(String username);
 
     ArrayList<Usuario> findAllBy();
@@ -22,6 +22,8 @@ public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
     Optional<Usuario> findUsuarioByUsername(String username);
 
     Page<Usuario> findAll(Pageable pageable);
+
+
 
 
 }
